# lib/units/api/swagger/api_v1.yaml
swagger: '2.0'
info:
  version: "2.3.0"
  title: Smartphone Test Farm
  description: Control and manages real Smartphone devices from browser and restful apis
  license:
    name: Apache-2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0'
  contact:
    name: STF Team
    email: contact@openstf.io
    url: 'http://openstf.io/'
basePath: /api/v1
schemes:
  - http
  - https
consumes:
  - application/json
  - application/octet-stream
produces:
  - application/json
...
paths:
  /user:
    x-swagger-router-controller: user
    get:
      summary: User Profile
      description: >-
        The User Profile endpoint returns information about current authorized
        user
      operationId: getUser
      tags:
        - user
      responses:
        '200':
          description: Current User Profile information
          schema:
            $ref: '#/definitions/UserResponse'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
        - accessTokenAuth: []
  /users:
    x-swagger-router-controller: user
    get:
      summary: Get All User List
      description: >-
        The User Profile endpoint returns information about current authorized
        user
      operationId: getUserList
      tags:
        - user
      responses:
        '200':
          description: Current User Profile information
          schema:
            $ref: '#/definitions/UserListResponse'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
        - accessTokenAuth: []
  /user/tags/:
    x-swagger-router-controller: user
    post:
      summary: Set User Tags
      description: >-
        The User Profile endpoint returns information about current authorized
        user
      operationId: setUserTags
      tags:
        - user
      parameters:
        - name: fields
          in: query
          description: >-
            Fields query parameter takes a comma seperated list of fields. Only
            listed field will be return in response
          required: false
          type: string
        - name: user
          in: body
          description: user
          required: true
          schema:
            $ref: '#/definitions/QueryUserTagsPayload'
      responses:
        '200':
          description: Current User Profile information
          schema:
            $ref: '#/definitions/UserListResponse'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
        - accessTokenAuth: []
  /servers:
    x-swagger-router-controller: user
    get:
      summary: Get All Server List
      description: >-
        The User Profile endpoint returns information about current authorized
        user
      operationId: getServerList
      tags:
        - user
      responses:
        '200':
          description: Current User Profile information
          schema:
            $ref: '#/definitions/ServerListResponse'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
        - accessTokenAuth: []
  /server/tags/:
    x-swagger-router-controller: user
    post:
      summary: Set Server Tags
      description: >-
        The User Profile endpoint returns information about current authorized
        user
      operationId: setServerTags
      tags:
        - user
      parameters:
        - name: fields
          in: query
          description: >-
            Fields query parameter takes a comma seperated list of fields. Only
            listed field will be return in response
          required: false
          type: string
        - name: server
          in: body
          description: server
          required: true
          schema:
            $ref: '#/definitions/QueryServerTagsPayload'
      responses:
        '200':
          description: Current User Profile information
          schema:
            $ref: '#/definitions/ServerListResponse'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
        - accessTokenAuth: []
  /user/devices:
    x-swagger-router-controller: user
    get:
      summary: User Devices
      description: >-
        The User Devices endpoint returns device list owner by current
        authorized user
      operationId: getUserDevices
      tags:
        - user
      parameters:
        - name: fields
          in: query
          description: >-
            Fields query parameter takes a comma seperated list of fields. Only
            listed field will be return in response
          required: false
          type: string
      responses:
        '200':
          description: Current User Devices List
          schema:
            $ref: '#/definitions/DeviceListResponse'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
        - accessTokenAuth: []
    post:
      summary: Add a device to a user
      description: The User Devices endpoint will request stf server for a new device.
      operationId: addUserDevice
      tags:
        - user
      parameters:
        - name: device
          in: body
          description: Device to add
          required: true
          schema:
            $ref: '#/definitions/AddUserDevicePayload'
      responses:
        '200':
          description: Add User Device Status
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
        - accessTokenAuth: []
  '/user/devices/{serial}':
    x-swagger-router-controller: user
    get:
      summary: User Device
      description: The devices enpoint return information about device owned by user
      operationId: getUserDeviceBySerial
      tags:
        - user
      parameters:
        - name: serial
          in: path
          description: Device Serial
          required: true
          type: string
        - name: fields
          in: query
          description: >-
            Fields query parameter takes a comma seperated list of fields. Only
            listed field will be return in response
          required: false
          type: string
      responses:
        '200':
          description: Device Information owned by user
          schema:
            $ref: '#/definitions/DeviceResponse'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
        - accessTokenAuth: []
    delete:
      summary: Delete User Device
      description: >-
        The User Devices endpoint will request for device release from stf
        server. It will return request accepted if device is being used by
        current user
      operationId: deleteUserDeviceBySerial
      tags:
        - user
      parameters:
        - name: serial
          in: path
          description: Device Serial
          required: true
          type: string
      responses:
        '200':
          description: Delete User Device Status
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
        - accessTokenAuth: []
  '/user/devices/{serial}/remoteConnect':
    x-swagger-router-controller: user
    post:
      summary: Remote Connect
      description: The device connect endpoint will request stf server to connect remotely
      operationId: remoteConnectUserDeviceBySerial
      tags:
        - user
      parameters:
        - name: serial
          in: path
          description: Device Serial
          required: true
          type: string
      responses:
        '200':
          description: Remote Connect User Device Request Status
          schema:
            $ref: '#/definitions/RemoteConnectUserDeviceResponse'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
        - accessTokenAuth: []
    delete:
      summary: Remote Disconnect
      description: >-
        The device connect endpoint will request stf server to disconnect
        remotely
      operationId: remoteDisconnectUserDeviceBySerial
      tags:
        - user
      parameters:
        - name: serial
          in: path
          description: Device Serial
          required: true
          type: string
      responses:
        '200':
          description: Remote Disonnect User Device Request Status
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
        - accessTokenAuth: []
  /user/accessTokens:
    x-swagger-router-controller: user
    get:
      summary: Access Tokens
      description: >-
        The Access Tokens endpoints returns titles of all the valid access
        tokens
      operationId: getUserAccessTokens
      tags:
        - user
      responses:
        '200':
          description: Access Tokens titles
          schema:
            $ref: '#/definitions/AccessTokensResponse'
        default:
          description: Unexpected Error
          schema:
            $ref: '#/definitions/ErrorResponse'
      security:
        - accessTokenAuth: []
  /user/adbkeys:
    x-swagger-router-controller: user
    post:
      summary: Adb public keys
      description: Add adb public key for current user
      operationId: addAdbPublicKey
      tags:
        - user
      parameters:
      - name: adb
        in: body
        schema:
            type: object
            required:
              - adb_key
            properties:
              adb_key:
                type: string
                description: adb public key (~/.android/id_rsa.pub)
              title:
                type: string
                description: By default will be extracted from public key
      responses:
        "200":
          description: Adb key added successfully
        default:
          description: Unexpected Error
          schema:
            $ref: "#/definitions/ErrorResponse"
      security:
        - accessTokenAuth: []
...
definitions:
  UserResponse:
    required:
      - user
    properties:
      user:
        type: object
  UserListResponse:
    required:
      - users
    properties:
      users:
        type: object
  ServerListResponse:
    required:
      - servers
    properties:
      servers:
        type: object
  AccessTokensResponse:
    required:
      - tokens
    properties:
      tokens:
        type: array
        items:
          type: string
  DeviceListResponse:
    required:
      - devices
    properties:
      devices:
        type: array
        items:
          type: object
  DeviceResponse:
    required:
      - device
    properties:
      device:
        type: object
  RemoteConnectUserDeviceResponse:
    required:
      - remoteConnectUrl
      - serial
    properties:
      remoteConnectUrl:
        type: string
      serial:
        type: string
...
  QueryUserTagsPayload:
    description: payload object for query testings
    required:
      - email
      - tags
    properties:
      email:
        description: user email
        type: string
      tags:
        description: user tags
        type: string
  QueryServerTagsPayload:
    description: payload object for query testings
    required:
      - name
    properties:
      name:
        description: provider name
        type: string
      note:
        description: server note
        type: string
      tags:
        description: user tags
        type: string
...
  AddUserDevicePayload:
    description: payload object for adding device to user
    required:
      - serial
    properties:
      serial:
        description: Device Serial
        type: string
      timeout:
        description: >-
          Device timeout in ms. If device is kept idle for this period, it will
          be automatically disconnected. Default is provider group timeout
        type: integer
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
securityDefinitions:
  accessTokenAuth:
    type: apiKey
    name: authorization
    in: header